+++ src/credit_management/db/base.py
@@ -40,20 +40,16 @@ class BaseDBManager(ABC):
 
     # User operations
     @abstractmethod
-    async def add_user(self, user: UserAccount) -> UserAccount:
-        ...
+    async def add_user(self, user: UserAccount) -> UserAccount: ...
 
     @abstractmethod
-    async def get_user(self, user_id: str) -> Optional[UserAccount]:
-        ...
+    async def get_user(self, user_id: str) -> Optional[UserAccount]: ...
 
     @abstractmethod
-    async def update_user(self, user: UserAccount) -> UserAccount:
-        ...
+    async def update_user(self, user: UserAccount) -> UserAccount: ...
 
     @abstractmethod
-    async def get_user_credits(self, user_id: str) -> int:
-        ...
+    async def get_user_credits(self, user_id: str) -> float: ...
 
     @abstractmethod
     async def get_user_credits_info(self, user_id: str) -> UserCreditInfo:
@@ -65,44 +61,31 @@ class BaseDBManager(ABC):
 
     # Transaction / ledger operations
     @abstractmethod
-    async def add_transaction(self, tx: Transaction) -> Transaction:
-        ...
+    async def add_transaction(self, tx: Transaction) -> Transaction: ...
 
     @abstractmethod
-    async def get_transaction(self, transaction_id: str) -> Optional[Transaction]:
-        ...
+    async def get_transaction(self, transaction_id: str) -> Optional[Transaction]: ...
 
     @abstractmethod
-    async def get_transactions(self, user_id: str) -> Iterable[Transaction]:
-        ...
+    async def get_transactions(self, user_id: str) -> Iterable[Transaction]: ...
 
     # Credit expiry / reservation
     @abstractmethod
-    async def add_credit_expiry_record(
-        self, record: CreditExpiryRecord
-    ) -> CreditExpiryRecord:
-        ...
+    async def add_credit_expiry_record(self, record: CreditExpiryRecord) -> CreditExpiryRecord: ...
 
     @abstractmethod
-    async def get_credit_expiry_history(
-        self, user_id: str
-    ) -> Iterable[CreditExpiryRecord]:
-        ...
+    async def get_credit_expiry_history(self, user_id: str) -> Iterable[CreditExpiryRecord]: ...
 
     @abstractmethod
-    async def add_reserved_credits(
-        self, reserved: ReservedCredits
-    ) -> ReservedCredits:
-        ...
+    async def add_reserved_credits(self, reserved: ReservedCredits) -> ReservedCredits: ...
 
     @abstractmethod
     async def get_reserved_credits_for_subscription_plan(
         self, subscription_plan_id: str
-    ) -> Iterable[ReservedCredits]:
-        ...
+    ) -> Iterable[ReservedCredits]: ...
 
     @abstractmethod
-    async def get_reserved_credits_for_user(self, user_id: str) -> int:
+    async def get_reserved_credits_for_user(self, user_id: str) -> float:
         """
         Sum of credits currently reserved for this user (not yet committed or released).
         Used to compute available balance = get_user_credits - get_reserved_credits_for_user.
@@ -111,60 +94,36 @@ class BaseDBManager(ABC):
 
     # Subscription operations
     @abstractmethod
-    async def add_subscription_plan(
-        self, plan: SubscriptionPlan
-    ) -> SubscriptionPlan:
-        ...
+    async def add_subscription_plan(self, plan: SubscriptionPlan) -> SubscriptionPlan: ...
 
     @abstractmethod
-    async def update_subscription_plan(
-        self, plan: SubscriptionPlan
-    ) -> SubscriptionPlan:
-        ...
+    async def update_subscription_plan(self, plan: SubscriptionPlan) -> SubscriptionPlan: ...
 
     @abstractmethod
-    async def delete_subscription_plan(self, plan_id: str) -> None:
-        ...
+    async def delete_subscription_plan(self, plan_id: str) -> None: ...
 
     @abstractmethod
-    async def get_subscription_plan(self, plan_id: str) -> Optional[SubscriptionPlan]:
-        ...
+    async def get_subscription_plan(self, plan_id: str) -> Optional[SubscriptionPlan]: ...
 
     @abstractmethod
-    async def get_all_subscription_plans(self) -> Iterable[SubscriptionPlan]:
-        ...
+    async def get_all_subscription_plans(self) -> Iterable[SubscriptionPlan]: ...
 
     @abstractmethod
-    async def add_user_subscription(
-        self, user_subscription: UserSubscription
-    ) -> UserSubscription:
-        ...
+    async def add_user_subscription(self, user_subscription: UserSubscription) -> UserSubscription: ...
 
     @abstractmethod
-    async def get_user_subscription_plan(
-        self, user_id: str
-    ) -> Optional[UserSubscription]:
-        ...
+    async def get_user_subscription_plan(self, user_id: str) -> Optional[UserSubscription]: ...
 
     @abstractmethod
-    async def update_user_subscription_plan(
-        self, user_subscription: UserSubscription
-    ) -> UserSubscription:
-        ...
+    async def update_user_subscription_plan(self, user_subscription: UserSubscription) -> UserSubscription: ...
 
     @abstractmethod
-    async def delete_user_subscription_plan(self, user_id: str) -> None:
-        ...
+    async def delete_user_subscription_plan(self, user_id: str) -> None: ...
 
     # Notifications
     @abstractmethod
-    async def add_notification_event(
-        self, notification: NotificationEvent
-    ) -> NotificationEvent:
-        ...
+    async def add_notification_event(self, notification: NotificationEvent) -> NotificationEvent: ...
 
     # Ledger
     @abstractmethod
-    async def add_ledger_entry(self, entry: LedgerEntry) -> LedgerEntry:
-        ...
-
+    async def add_ledger_entry(self, entry: LedgerEntry) -> LedgerEntry: ...
