+++ src/credit_management/models/api_models.py
@@ -3,25 +3,25 @@ from pydantic import BaseModel
 
 class AddCreditsRequest(BaseModel):
     user_id: str
-    amount: int
+    amount: float
     description: str | None = None
 
 
 class DeductCreditsRequest(BaseModel):
     user_id: str
-    amount: int
+    amount: float
     description: str | None = None
 
 
 class CreditBalanceResponse(BaseModel):
     user_id: str
-    credits: int
+    credits: float
 
 
 class SubscriptionPlanRequest(BaseModel):
     name: str
     description: str | None = None
-    credit_limit: int
+    credit_limit: float
     price: float
     billing_period: str
     validity_days: int
@@ -30,7 +30,7 @@ class SubscriptionPlanRequest(BaseModel):
 class SubscriptionPlanResponse(BaseModel):
     id: str
     name: str
-    credit_limit: int
+    credit_limit: float
     price: float
     billing_period: str
     validity_days: int
